requirements.txt
streamlit
pypdf
langchain-google-genai
python-dotenv
pandas
